+++ packages/create-qwik/package.json
@@ -28,7 +28,6 @@
     "url": "https://github.com/BuilderIO/qwik/issues"
   },
   "devDependencies": {
-    "@clack/prompts": "^0.5.0",
     "kleur": "4.1.5"
   },
   "engines": {
