+++ packages/qwik/src/cli/run.ts
@@ -2,10 +2,14 @@
 import color from 'kleur';
 import { AppCommand } from './utils/app-command';
 import { runAddCommand } from './add/run-add-command';
-import { panic, pmRunCmd } from './utils/utils';
+import { note, panic, pmRunCmd, printHeader } from './utils/utils';
 import { runBuildCommand } from './build/run-build-command';
+import { intro, outro } from '@clack/prompts';
 
 export async function runCli() {
+  console.clear();
+  printHeader();
+
   try {
     const app = new AppCommand({
       rootDir: '',
@@ -48,6 +52,20 @@ async function runCommand(app: AppCommand) {
 
 async function printHelp() {
   const pmRun = pmRunCmd();
+
+  intro(color.bgMagenta().white(` Qwik Help `));
+
+  note(
+    `
+${pmRun} qwik ${color.cyan(`add`)}            ${color.dim(`Add an integration to this app`)}
+${pmRun} qwik ${color.cyan(`build`)}          ${color.dim(`Parallelize builds and type checking`)}
+${pmRun} qwik ${color.cyan(`build preview`)}  ${color.dim(
+      `Same as "build", but for preview server`
+    )}
+  `,
+    'Available commands'
+  );
+
   console.log(``);
   console.log(color.bgMagenta(` Qwik Help `));
   console.log(``);
